feat(wallet-dashboard): add missing link for AllianceNo2
