+++ apps/wallet-dashboard/providers/index.ts
@@ -1,0 +1,4 @@
+// Copyright (c) 2024 IOTA Stiftung
+// SPDX-License-Identifier: Apache-2.0
+
+export * from './AppProviders';
